Standard ML day 19
It's not slow.. but it can be significantly faster if we left factor to
avoid expanding a big chunk yet again. Consider this case:

    1: 2 "a" | 2 "b"

Here you'll have to match 2 twice, which again may have matches that has
to be matched twice, and so on. I guess the pathological case is
O(exponential) to the number of rules or something?